refactor: move WarmUp out of udp.go into warmup.go
WarmUp only jitters and delegates to the transport, it touches no UDP
state. It lived in udp.go because the first implementation warmed the
path by opening a UDP exchange to the proxied DNS server; when that was
replaced by a /v3/probe request the placement stopped matching the code
but stayed behind.

Move the method verbatim into client/proxy/warmup.go, alongside the
warmup_test.go that already covers it, and drop the now unused
math/rand/v2 import from udp.go. No behavior change.
